pyproject.toml
[tool.poetry]
name = "faq-chatbot"
version = "0.1.0"
description = ""
authors = ["yuhyeongchan <gudcks305@gmail.com>"]
readme = "README.md"
packages = [{include = "faq_chatbot"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
openai = "^1.57.0"
pandas = "^2.2.3"
jupyter = "^1.1.1"
pymilvus = "^2.4.6"
grpcio = ">=1.49.1,<=1.67.1"
python-dotenv = "^1.0.1"
pynest-api = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
